nlp/generative-ai/generator/requirements.txt
PyPDF2 
python-docx